l2fwd: store Mpps samples as double instead of uint32_t
samples was declared uint32_t* but print_stats() stores a double Mpps
value into it, so every per-second sample was silently floor()ed. min and
max were then read back as uint64_t and printed with %lu, so the reported
Minimum/Maximum/Average were all computed from truncated integers.

The bias is a fixed ~0.5 Mpps, which makes it invisible where throughput
is high and severe where it is low: -0.3% at 93 Mpps but -4.6% at 17
Mpps. That is backwards from what you want, because the low-throughput
points are exactly where the unsaturated per-queue scaling slope is
measured.

The truncation also made the reported range useless. Sample 0 is always
cold -- 12-19% below steady state -- so it always set Minimum, meaning
the published min/max was a startup artifact rather than run-to-run
spread.

Verified against the old behaviour on the same configuration:

    before:  Minimum: 23      Maximum: 28      Average: 27.84
    after:   Minimum: 22.43   Maximum: 27.55   Average: 27.26

Correcting this made maglev's scaling series resolve into a straight
line -- 17.62, 34.99, 52.34, 69.59, 86.94, i.e. increments of 17.37,
17.35, 17.25, 17.35 -- which the truncated data had obscured.

Two related truncations remain and are deliberately not touched here:
the sample interval at main.c:205 and the RX batch size at main.c:196
are both uint64_t/uint64_t integer divisions. The batch one matters most,
since it reports a true 1.9 as 1.

+++ l2fwd/main.c
@@ -163,7 +163,7 @@ void print_port_stats(uint16_t portid) {
 }
 
 uint32_t SAMPLE_SIZE = 0;
-uint32_t *samples;
+double *samples;
 #define TOTAL_SAMPLES 32
 
 uint64_t total_packets_fwded_prev = 0;
@@ -226,8 +226,8 @@ static void print_final_stats() {
   }
 
   // Initialize min and max to the first element in the array
-  uint64_t min = samples[0];
-  uint64_t max = samples[0];
+  double min = samples[0];
+  double max = samples[0];
   double sum =
       0.0; // Use double to prevent integer overflow and get an accurate average
 
@@ -248,8 +248,8 @@ static void print_final_stats() {
   double avg = sum / SAMPLE_SIZE;
 
   // Print the results
-  printf("Minimum: %lu\n", min);
-  printf("Maximum: %lu\n", max);
+  printf("Minimum: %.2f\n", min);
+  printf("Maximum: %.2f\n", max);
   printf("Average: %.2f\n", avg);
 }
 
@@ -660,7 +660,7 @@ int main(int argc, char **argv) {
   else if (l2fwd_dramblast_enabled)
     dramblast_init();
 
-  samples = malloc(TOTAL_SAMPLES * sizeof(uint32_t));
+  samples = malloc(TOTAL_SAMPLES * sizeof(double));
   sleep(5);
 
   RTE_ETH_FOREACH_DEV(portid) {
